requirements.txt
Pillow==9.4.0
numpy==1.24.1
scipy==1.10.0
